+++ pkg/analysis_server/lib/src/session_logger/entry_keys.dart
@@ -1,0 +1,22 @@
+// Copyright (c) 2025, the Dart project authors. Please see the AUTHORS file
+// for details. All rights reserved. Use of this source code is governed by a
+// BSD-style license that can be found in the LICENSE file.
+
+/// The key used in an entry to access the command-line arguments associated
+/// with the entry.
+const String argList = 'argList';
+
+/// The key used in an entry to access the kind of the entry.
+const String kind = 'kind';
+
+/// The key used in an entry to access the message associatd with the entry.
+const String message = 'message';
+
+/// The key used in an entry to access the receiver of a message.
+const String receiver = 'receiver';
+
+/// The key used in an entry to access the sender of a message.
+const String sender = 'sender';
+
+/// The key used in an entry to access the sender of a message.
+const String time = 'time';