+++ pkg/analyzer_plugin/lib/utilities/change_builder/change_builder_dart.dart
@@ -13,10 +13,6 @@ import 'package:analyzer/dart/element/type_system.dart';
 import 'package:analyzer/source/source_range.dart';
 import 'package:analyzer_plugin/utilities/change_builder/change_builder_core.dart';
 
-const _deprecationMessageMethodBeingCopied =
-    'Use `typeParametersInScope` '
-    'instead. You can use `methodBeingCopied.typeParameters`.';
-
 /// The optional generator for prefix that should be used for new imports.
 typedef ImportPrefixGenerator = String Function(Uri);
 
@@ -33,12 +29,6 @@ abstract class DartEditBuilder implements EditBuilder {
   /// Checks whether the code for a type annotation for the given [type] can be
   /// written.
   ///
-  /// {@template methodBeingCopied}
-  /// If a [methodBeingCopied] is provided, then type parameters defined by that
-  /// method are assumed to be part of what is being written and hence valid
-  /// types.
-  /// {@endtemplate}
-  ///
   /// {@template typeParametersInScope}
   /// If [typeParametersInScope] is provided, then it will be used to resolve
   /// what are the valid type parameters to use.
@@ -47,8 +37,6 @@ abstract class DartEditBuilder implements EditBuilder {
   /// The logic is the same as the one used in [writeType].
   bool canWriteType(
     DartType? type, {
-    @Deprecated(_deprecationMessageMethodBeingCopied)
-    ExecutableElement? methodBeingCopied,
     List<TypeParameterElement>? typeParametersInScope,
   });
 
@@ -135,8 +123,6 @@ abstract class DartEditBuilder implements EditBuilder {
 
   /// Writes the code for a single parameter with the given [name].
   ///
-  /// {@macro methodBeingCopied}
-  ///
   /// {@macro typeParametersInScope}
   ///
   /// If a [type] is provided, then it will be used as the type of the
@@ -161,8 +147,6 @@ abstract class DartEditBuilder implements EditBuilder {
     String name, {
     bool isCovariant,
     bool isRequiredNamed,
-    @Deprecated(_deprecationMessageMethodBeingCopied)
-    ExecutableElement? methodBeingCopied,
     List<TypeParameterElement>? typeParametersInScope,
     String? nameGroupName,
     DartType? type,
@@ -173,18 +157,11 @@ abstract class DartEditBuilder implements EditBuilder {
   /// Writes the code for a list of [parameters], including the surrounding
   /// parentheses and default values (unless [includeDefaultValues] is `false`).
   ///
-  /// {@macro methodBeingCopied}
-  ///
   /// {@macro typeParametersInScope}
   ///
   /// If [requiredTypes] is `true`, then the types are always written.
   void writeFormalParameters(
     Iterable<FormalParameterElement> parameters, {
-    @Deprecated(
-      '$_deprecationMessageMethodBeingCopied And for the group prefix, '
-      'inform `groupNamePrefix`.',
-    )
-    ExecutableElement? methodBeingCopied,
     List<TypeParameterElement>? typeParametersInScope,
     String? groupNamePrefix,
     bool fillParameterNames = true,
@@ -298,8 +275,6 @@ abstract class DartEditBuilder implements EditBuilder {
 
   /// Writes the code for a single parameter with the given [name].
   ///
-  /// {@macro methodBeingCopied}
-  ///
   /// {@macro typeParametersInScope}
   ///
   /// If a [type] is provided, then it will be used as the type of the
@@ -322,8 +297,6 @@ abstract class DartEditBuilder implements EditBuilder {
     String name, {
     bool isCovariant,
     bool isRequiredNamed,
-    @Deprecated(_deprecationMessageMethodBeingCopied)
-    ExecutableElement? methodBeingCopied,
     List<TypeParameterElement>? typeParametersInScope,
     String? nameGroupName,
     DartType? type,
@@ -344,8 +317,6 @@ abstract class DartEditBuilder implements EditBuilder {
     Expression argument,
     int index,
     Set<String> usedNames, {
-    @Deprecated(_deprecationMessageMethodBeingCopied)
-    ExecutableElement? methodBeingCopied,
     List<TypeParameterElement>? typeParametersInScope,
   });
 
@@ -357,8 +328,6 @@ abstract class DartEditBuilder implements EditBuilder {
   /// {@macro typeParametersInScope}
   void writeParametersMatchingArguments(
     ArgumentList arguments, {
-    @Deprecated(_deprecationMessageMethodBeingCopied)
-    ExecutableElement? methodBeingCopied,
     List<TypeParameterElement>? typeParametersInScope,
   });
 
@@ -387,8 +356,6 @@ abstract class DartEditBuilder implements EditBuilder {
     String? nameGroupName,
     DartType? parameterType,
     String? parameterTypeGroupName,
-    @Deprecated(_deprecationMessageMethodBeingCopied)
-    ExecutableElement? methodBeingCopied,
     List<TypeParameterElement>? typeParametersInScope,
   });
 
@@ -406,8 +373,6 @@ abstract class DartEditBuilder implements EditBuilder {
   /// [addSupertypeProposals] is `true`, then all of the supertypes of the
   /// [type] will be added as suggestions for alternatives to the type name.
   ///
-  /// {@macro methodBeingCopied}
-  ///
   /// {@macro typeParametersInScope}
   ///
   /// Returns `true` if any text was written.
@@ -415,8 +380,6 @@ abstract class DartEditBuilder implements EditBuilder {
     DartType? type, {
     bool addSupertypeProposals = false,
     String? groupName,
-    @Deprecated(_deprecationMessageMethodBeingCopied)
-    ExecutableElement? methodBeingCopied,
     List<TypeParameterElement>? typeParametersInScope,
     bool required = false,
     bool shouldWriteDynamic = false,
@@ -426,26 +389,18 @@ abstract class DartEditBuilder implements EditBuilder {
   ///
   /// The enclosing angle brackets are not automatically written.
   ///
-  /// {@macro methodBeingCopied}
-  ///
   /// {@macro typeParametersInScope}
   void writeTypeParameter(
     TypeParameterElement typeParameter, {
-    @Deprecated(_deprecationMessageMethodBeingCopied)
-    ExecutableElement? methodBeingCopied,
     List<TypeParameterElement>? typeParametersInScope,
   });
 
   /// Writes the code to declare the given list of [typeParameters]. The
   /// enclosing angle brackets are automatically written.
   ///
-  /// {@macro methodBeingCopied}
-  ///
   /// {@macro typeParametersInScope}
   void writeTypeParameters(
     List<TypeParameterElement> typeParameters, {
-    @Deprecated(_deprecationMessageMethodBeingCopied)
-    ExecutableElement? methodBeingCopied,
     List<TypeParameterElement>? typeParametersInScope,
   });
 
@@ -497,13 +452,9 @@ abstract class DartFileEditBuilder implements FileEditBuilder {
   /// Checks whether the code for a type annotation for the given [type] can be
   /// written.
   ///
-  /// {@macro methodBeingCopied}
-  ///
   /// {@macro typeParametersInScope}
   bool canWriteType(
     DartType? type, {
-    @Deprecated(_deprecationMessageMethodBeingCopied)
-    ExecutableElement? methodBeingCopied,
     List<TypeParameterElement>? typeParametersInScope,
   });