+++ pkg/analyzer_plugin/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.14.0-dev
+
+- Require version `9.0.1` of the `analyzer` package.
+- `AssistContributorMixin` is now a mixin.
+- Remove deprecated `methodBeingCopied` parameters on various methods.
+
 ## 0.13.11
 
 - Require version `9.0.0` of the `analyzer` package.